* YAM_UT.c: when closing a temp file which was never opened before YAM did   nevertheless try to delete this file which of course must fail. As a   consequence this file was treated as a zombie file furtheron. When all this   happened while YAM was shutting down adding the file to the zombie file list   caused a crash, because the timer which then got restarted had been deleted   already. Now only existing files will be treated as zombie files. This   closes #242.

+++ src/YAM_UT.c
@@ -1500,9 +1500,12 @@ void CloseTempFile(struct TempFile *tf)
     if(tf->FP != NULL)
       fclose(tf->FP);
 
-    D(DBF_UTIL, "DeleteTempFile: %s", tf->Filename);
-    if(DeleteFile(tf->Filename) == 0)
-      AddZombieFile(tf->Filename);
+    if(FileExists(tf->Filename) == TRUE)
+    {
+      D(DBF_UTIL, "DeleteTempFile: '%s'", tf->Filename);
+      if(DeleteFile(tf->Filename) == 0)
+        AddZombieFile(tf->Filename);
+    }
 
     free(tf);
   }
